requirements.txt
streamlit>=1.30.0   # For the Dashboard UI
pandas>=2.0.0       # For reading the SQLite database efficiently
fastapi
uvicorn
requests
graphviz
openai
PyYAML


#backend
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
duckdb>=0.9.0
requests>=2.31.0
pydantic>=2.0.0